+++ packages/expo-av/src/AV.ts
@@ -1,7 +1,14 @@
 import { Asset } from 'expo-asset';
 import { Platform } from 'expo-modules-core';
 
-import ExponentAV from './ExponentAV';
+import {
+  AVPlaybackSource,
+  AVPlaybackNativeSource,
+  AVPlaybackStatus,
+  AVPlaybackStatusToSet,
+  PitchCorrectionQuality,
+} from './AV.types';
+
 // TODO add:
 //  disableFocusOnAndroid
 //  audio routes (at least did become noisy on android)
@@ -10,78 +17,6 @@ import ExponentAV from './ExponentAV';
 //  API to explicitly request audio focus / session
 //  API to select stream type on Android
 //  subtitles API
-
-export enum PitchCorrectionQuality {
-  Low = ExponentAV && ExponentAV.Qualities && ExponentAV.Qualities.Low,
-  Medium = ExponentAV && ExponentAV.Qualities && ExponentAV.Qualities.Medium,
-  High = ExponentAV && ExponentAV.Qualities && ExponentAV.Qualities.High,
-}
-
-export type AVPlaybackSource =
-  | number
-  | {
-      uri: string;
-      overrideFileExtensionAndroid?: string;
-      headers?: { [fieldName: string]: string };
-    }
-  | Asset;
-
-export type AVPlaybackNativeSource = {
-  uri: string;
-  overridingExtension?: string | null;
-  headers?: { [fieldName: string]: string };
-};
-
-export type AVMetadata = {
-  title?: string;
-};
-
-export type AVPlaybackStatus =
-  | {
-      isLoaded: false;
-      androidImplementation?: string;
-      error?: string; // populated exactly once when an error forces the object to unload
-    }
-  | {
-      isLoaded: true;
-      androidImplementation?: string;
-
-      uri: string;
-
-      progressUpdateIntervalMillis: number;
-      durationMillis?: number;
-      positionMillis: number;
-      playableDurationMillis?: number;
-      seekMillisToleranceBefore?: number;
-      seekMillisToleranceAfter?: number;
-
-      shouldPlay: boolean;
-      isPlaying: boolean;
-      isBuffering: boolean;
-
-      rate: number;
-      shouldCorrectPitch: boolean;
-      volume: number;
-      isMuted: boolean;
-      isLooping: boolean;
-
-      didJustFinish: boolean; // true exactly once when the track plays to finish
-    };
-
-export type AVPlaybackStatusToSet = {
-  androidImplementation?: string;
-  progressUpdateIntervalMillis?: number;
-  positionMillis?: number;
-  seekMillisToleranceBefore?: number;
-  seekMillisToleranceAfter?: number;
-  shouldPlay?: boolean;
-  rate?: number;
-  shouldCorrectPitch?: boolean;
-  volume?: number;
-  isMuted?: boolean;
-  isLooping?: boolean;
-  pitchCorrectionQuality?: PitchCorrectionQuality;
-};
 
 export const _DEFAULT_PROGRESS_UPDATE_INTERVAL_MILLIS: number = 500;
 export const _DEFAULT_INITIAL_PLAYBACK_STATUS: AVPlaybackStatusToSet = {
@@ -326,3 +261,5 @@ export const PlaybackMixin = {
     return (this as any as Playback).setStatusAsync({ progressUpdateIntervalMillis });
   },
 };
+
+export * from './AV.types';
