+++ packages/expo-av/src/ExponentAV.web.ts
@@ -1,7 +1,8 @@
 import { PermissionResponse, PermissionStatus, SyntheticPlatformEmitter } from 'expo-modules-core';
 
-import { AVPlaybackNativeSource, AVPlaybackStatus, AVPlaybackStatusToSet } from './AV';
-import { RecordingStatus, RECORDING_OPTIONS_PRESET_HIGH_QUALITY } from './Audio/Recording';
+import type { AVPlaybackNativeSource, AVPlaybackStatus, AVPlaybackStatusToSet } from './AV.types';
+import type { RecordingStatus } from './Audio/Recording.types';
+import { RECORDING_OPTIONS_PRESET_HIGH_QUALITY } from './Audio/RecordingConstants';
 
 async function getPermissionWithQueryAsync(name: PermissionName): Promise<PermissionStatus | null> {
   if (!navigator || !navigator.permissions || !navigator.permissions.query) return null;
